roadbuddy_winning_arch/requirements.txt
# RoadBuddy Winning Architecture - Requirements
# Python 3.10+

# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Vision Models
timm>=0.9.12
einops>=0.7.0

# Object Detection
ultralytics>=8.1.0
supervision>=0.18.0

# OCR Ensemble
vietocr>=0.3.11
paddleocr>=2.7.0
easyocr>=1.7.0

# NLP & Embeddings
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4

# Video Processing
opencv-python>=4.9.0
av>=11.0.0
decord>=0.6.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.4
scikit-learn>=1.3.2
pillow>=10.1.0

# Configuration & Logging
pyyaml>=6.0.1
omegaconf>=2.3.0
tensorboard>=2.15.1
wandb>=0.16.2
tqdm>=4.66.1
loguru>=0.7.2

# Utilities
matplotlib>=3.8.2
seaborn>=0.13.0
jupyter>=1.0.0
ipywidgets>=8.1.1

# Testing & Quality
pytest>=7.4.3
black>=23.12.1
flake8>=7.0.0
isort>=5.13.2

# API & Deployment (Optional)
fastapi>=0.108.0
uvicorn>=0.25.0
gradio>=4.13.0
